Release 0 | Beta 0 | Development 2 | Experimental 2
* Added Object3DBuilder
* Added Quad3DTest
* Added OpenGLUtils

* Changed the type of Settings.Window.Width, and
  Settings.Window.Height to float instead of
  int

* Window now sets the size of the window after it
  has been created so fullscreen windows will also
  set the width and height properly

+++ Andor/src/org/andor/core/Settings.java
@@ -25,8 +25,8 @@ public class Settings {
 		/* The title of the window */
 		public static String Title = "Andor";
 		/* The width and height of the window */
-		public static int Width = 800;
-		public static int Height = 600;
+		public static float Width = 800;
+		public static float Height = 600;
 		/* The boolean that states whether the window is fullscreen */
 		public static boolean Fullscreen = false;
 	}
